zadanie 2

+++ sem3/p1.c
@@ -3,8 +3,9 @@
 
 int main (int argc, char *argv[])
 {
+        
     float v;
-    int a1;
+    float a1;
     float a2;
     a1 = atof(argv[1]);
     a2 = atof(argv[2]);
